put ResourceAccess fake into own package
